Rename project organizer file
